Initial commit of a big rewrite.
Move away from Run in favour of XMonad.Util.Commands. This should
allow for a general expression of commands which should in turn
simplify much of the logic I need to express complex ideas with
workspaces.

spawnIn needs to be implemented, its currently stubbed out with a
undefined. Support is going to need yet another overhaul with how
I handle commands yet again.