Rethink Commands again, interface changes.
Command String now works like safe Spawn with no arguments. This is
much more common behaviour and more convenient. If I want spawn to
run shell script I have to explicitly wrap it with Shell. This, to me,
is a lot smarter and safer. One interface for external commands and
intention is explicit.